% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqUtilities.R
\name{createDEGGraphs}
\alias{createDEGGraphs}
\title{a histogram of FDRs. Input the return value from topTable (limma) with column names
unchanged.}
\usage{
createDEGGraphs(dat, title = NULL, fdrCut = 0.2)
}
\arguments{
\item{dat}{data.frame Return data.frame from topTable function in limma}

\item{title}{character string for title of graph.}

\item{fdrCut}{Significance cut-off value for FDR. Genes who pass cut-off will be 
shown in red in volcano plot.}
}
\value{
list with three ggplot2 elements - p-value histogram, volcano plot, and FDR
       histogram.
}
\description{
a histogram of FDRs. Input the return value from topTable (limma) with column names
unchanged.
}
