% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqUtilities.R
\name{headerKable}
\alias{headerKable}
\title{Create a nicer kable table.}
\usage{
headerKable(tabley, labels = NULL, levels = NULL, title = NULL,
  align = "left", escapeMe = FALSE, digitts = 3)
}
\arguments{
\item{tabley}{data.frame of sample sizes. Each column is a level and each row
is a level of the reference factor. The first column of the tabley will be total
and the remaining columns will be the levels.}

\item{labels}{vector of factor labels which will be the top level header.}

\item{levels}{vector of number of levels for each factor in 'labels'}

\item{title}{string for caption of table}

\item{align}{parameter for 'position' argument for kable_styling. default: left}

\item{escapeMe}{escape parameter for kable which is logical to escape special
symbols}

\item{digitts}{set number of significant digits to be displayed}
}
\value{
kable table
}
\description{
Creates an html kable table with additional parameters, including a
multi-level column label if desired. This is designed to be used on a table
created by \link{countSampleSizes}. If 'levels' or 'labels' are NULL then
no header will be produced.
}
