% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqUtilities.R
\name{outputDEG}
\alias{outputDEG}
\title{Create markdown string showing deg results.}
\usage{
outputDEG(deg, graphs, doSlides = TRUE, onlySigs = FALSE,
  cutOnPval = TRUE, FDRCut = 0.2, pvalCut = 0.01, header = "##")
}
\arguments{
\item{deg}{List of deg results from runDEG. Name of deg element will be used as title}

\item{graphs}{List of graphs from createDEGGraphs. Name of deg element will be used as title}

\item{doSlides}{Boolean: If TRUE then output markdown slides else produce standard markdown.}

\item{onlySigs}{Boolean: if TRUE output only genesets that pass the 'FDRCut' threshold
(output message if no significant gene sets) else output all genesets.
Default is FALSE.}

\item{cutOnPval}{Boolean: if TRUE use pvalCut for threshold else use FDRCut}

\item{FDRCut}{FDR threshold to use to determine significance. Default is 0.2.}

\item{pvalCut}{P value threshold to use to determine significance. Default is
0.01}

\item{header}{Character string defining markdown header to ouput for each set
of results. This is only used if doSlides=FALSE.}
}
\value{
Vector of markdown strings
}
\description{
Input values for parameters 'deg' and 'graphs' must be named deg and graphs.
TODO: explore htmltools:knitr_methods to output datatables instead.
}
