% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqUtilities.R
\name{outputGSEA}
\alias{outputGSEA}
\title{Create markdown string showing gsea results.}
\usage{
outputGSEA(gsea, doSlides = TRUE, onlySigs = TRUE, FDRCut = 0.2,
  header = "##")
}
\arguments{
\item{gsea}{List of gsea results from runGSEA. Name of gsea element will be used as title}

\item{doSlides}{Boolean: If TRUE then output markdown slides else produce standard markdown.}

\item{onlySigs}{Boolean: if TRUE output only genesets that pass the 'FDRCut' threshold
(output message if no significant gene sets) else output all genesets. 
Default is FALSE.}

\item{FDRCut}{FDR threshold to use to determine significance. Default is 0.2.}

\item{header}{Character string defining markdown header to ouput for each set
of results. This is only used if doSlides=FALSE.}
}
\value{
Vector of markdown strings
}
\description{
Create markdown string showing gsea results.
}
