% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqUtilities.R
\docType{package}
\name{RNASeqUtilities}
\alias{RNASeqUtilities}
\alias{RNASeqUtilities-package}
\title{User defined functions for use in RNASeq analysis}
\description{
Standardise commonly used commands
}
\details{
Create commonly used functions to simplify RNASeq analysis markdown reports.
}
\author{
Carl Murie
}
