% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqUtilities.R
\name{runGSEA}
\alias{runGSEA}
\title{Apply GSEA to a voom normalized data set for a specific model and coefficients.}
\usage{
runGSEA(voomDat, geneIndices, coefs, labels = NULL, onlySigs = FALSE,
  FDRCut = 0.2)
}
\arguments{
\item{voomDat}{Eset of voom normalised data which must include the design. This is the
output of voom function in limma.}

\item{geneIndices}{List of gene indices corresponding to gene sets. It is created by
ids2indices function in limma.}

\item{coefs}{Vector of coefficients to test. They must map to a coefficient in the 
design matrix in 'voomDat'.}

\item{labels}{Vector of character strings that labels each coefficient being tested. This
allows for more human readable labels than the lmFit coefficients. Labels 
must map one to one to 'coefs'.}

\item{onlySigs}{Boolean: if TRUE return only genesets that pass the 'FDRCut' threshold
 (return -1 if no significant gene sets) else return all genesets. Default
is FALSE.}

\item{FDRCut}{FDR threshold to use to determine significance. Default is 0.2.}
}
\value{
List of camera results. Each element is the result for a single coefficient from
        'coefs'
}
\description{
Apply GSEA to a voom normalized data set for a specific model and coefficients.
}
